Bug fixing and code auditing
- Validate AntiForgeryTokens
- Instead of throwing exceptions when AntiForgeryTokens were missing or
invalid post an error message instead. This still needs some work.
- Shorten the AntiForgeryToken length
- Expose the creating of AntiForgeryTokens to the Controller.
- Change the special Html helper bundle that creates a login user form
to require an AntiForgeryToken be passed to it via it's constructor. See
comments in Aurora.Extra.cs for why.
- Add a new DebugMode variable to AspNetAdapter to identify if the
application's web.config is in debug mode vice if the assembly was
compiled in debug mode.
- Make sure all areas that look up .NET assemblies do so in a way that
they won't fail if dependent assemblies cannot be loaded.